clean モッククラスに専用のファイルを用意した
WeatherViewControllerMockとSuccessFetcherMock専用のファイルを用意した。FetcherMockはSuccessFetcherMockに名前を変更した。refs #33
